CLI Syntax #67
- refactoring - unimportant changes
- UT :+1:
