--- Opc.Ua.ModelCompiler/app.config
@@ -1,2 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<configuration></configuration>
